config.py
import os
from dotenv import load_dotenv

load_dotenv()

class Config:
    API_ID = int(os.environ.get("API_ID", 0))
    API_HASH = os.environ.get("API_HASH", "")
    BOT_TOKEN = os.environ.get("BOT_TOKEN", "")
    MONGO_URL = os.environ.get("MONGO_URL", "")
    OWNER_ID = int(os.environ.get("OWNER_ID", 0))
    LOG_CHANNEL = int(os.environ.get("LOG_CHANNEL", 0))
    START_PIC = os.environ.get("START_PIC", "https://telegra.ph/file/your-image.jpg")
    BOT_USERNAME = os.environ.get("BOT_USERNAME", "")
    KOYEB_APP_NAME = os.environ.get("KOYEB_APP_NAME", "")
    KOYEB_API_KEY = os.environ.get("KOYEB_API_KEY", "")
    WEB_PORT = int(os.environ.get("PORT", 8080))
